Moved Container.databaseConnection method to Container+DatabaseConnectable.swift file
